requirements.txt
python-dotenv==1.1.1
langchain-community==0.3.20
langchain==0.3.27
langchain-pinecone==0.2.12
huggingface-hub==0.35.2
pinecone==7.3.0
langchain-google-genai==2.1.12
google-genai==1.39.1
pydantic==2.11.9
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
pypdf==6.1.1
docx2txt==0.9
